+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
 - Add local permissions and a "Local Manager" role.
   Permissions : imio.smartweb.core.CanEditMinisiteLogo, imio.smartweb.core.CanManageSectionHTML
   [boulch]
+- Handle inline SVG images for portal logo and minisite logo
+  [laulaz]
 
 - Add show_items_lead_image attributes on files section.
   Add no-image css class in table template when there is no image to display